+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Added
+* `sql` input `params` for placeholders
+
 ### Removed
 * `dense` ordering moved to `first`
 * `GraphQL.federated` replaced with class-based approach